manicomio: move layout + add description + add description + forced to fix typescript errors
