+++ src/context/layout.tsx
@@ -2,12 +2,9 @@ import React from "react";
 
 export type Props = {
   isSidebarOpen: boolean;
-  isCourseMenuOpen: boolean;
   isBannerOpen: boolean;
   setIsBannerOpen: React.Dispatch<React.SetStateAction<boolean>>;
   toggleSidebar: () => void;
-  toggleCourseMenu: () => void;
-  handleClickAway: () => void;
 };
 
 const LayoutContext = React.createContext<Props | null>(null);
@@ -18,8 +15,7 @@ const LayoutProvider = ({
   children: React.ReactChild | React.ReactChild[];
 }) => {
   const [isSidebarOpen, setIsSidebarOpen] = React.useState<boolean>(false);
-  const [isCourseMenuOpen, setIsCourseMenuOpen] =
-    React.useState<boolean>(false);
+
   const [isBannerOpen, setIsBannerOpen] = React.useState<boolean>(true);
 
   const toggleSidebar = React.useCallback(() => {
@@ -33,25 +29,14 @@ const LayoutProvider = ({
     });
   }, []);
 
-  const toggleCourseMenu = React.useCallback(() => {
-    setIsCourseMenuOpen((old) => !old);
-  }, []);
-
-  const handleClickAway = React.useCallback(() => {
-    setIsCourseMenuOpen(false);
-  }, []);
-
   const contextValue = React.useMemo(() => {
     return {
-      isCourseMenuOpen,
       isSidebarOpen,
       isBannerOpen,
       setIsBannerOpen,
-      toggleCourseMenu,
       toggleSidebar,
-      handleClickAway,
     };
-  }, [isCourseMenuOpen, isSidebarOpen, isBannerOpen]);
+  }, [isSidebarOpen, isBannerOpen]);
 
   return (
     <LayoutContext.Provider value={contextValue}>
