+++ src/styles/global.css
@@ -13,16 +13,19 @@
   height: 7px;
   background: white;
 }
+
 ::-o-scrollbar {
   width: 7px;
   height: 7px;
   background: white;
 }
+
 ::-moz-scrollbar {
   width: 7px;
   height: 7px;
   background: white;
 }
+
 ::-webkit-scrollbar-thumb {
   width: 7px;
   height: 7px;
@@ -36,6 +39,7 @@
   background: var(--purple-200);
   border-radius: 4px;
 }
+
 ::-moz-scrollbar-thumb {
   width: 7px;
   height: 7px;
@@ -61,7 +65,7 @@
   --primary-100: rgba(98, 0, 238, 0.05);
   --secondary-main: #E7A7FF;
   --tertiary-main: #F3DFFB;
-  --optional--main: #6200EE; 
+  --optional--main: #6200EE;
 
   /* Global */
   --letter-spacing: 0.08rem;
@@ -77,29 +81,29 @@
   /*
   Official Theme
   */
-      --gray-300: #E4E7EC;
-      --gray-400: #D0D5DD;
-      --gray-500: #6C757D;
-      --gray-600: #4B4B4B;
-      --gray-700: #343A40;
-      --gray-800: #000000;
+  --gray-300: #E4E7EC;
+  --gray-400: #D0D5DD;
+  --gray-500: #6C757D;
+  --gray-600: #4B4B4B;
+  --gray-700: #343A40;
+  --gray-800: #000000;
 
-      --purple-alpha-100: rgba(98, 0, 238, 0.025);
-      --purple-alpha-200: rgba(98, 0, 238, 0.05);
-      --purple-200: #E9E3FF;
-      --purple-300: #E7A7FF;
-      --purple-400: #8769FE;
-      --purple-500: #A435F0;
-      --purple-800: #341268;
-      --purple-gradient-200: linear-gradient(180deg, #E9E3FF, white );
-      --purple-gradient-800:: linear-gradient(180deg, #341268, #8769FE);
+  --purple-alpha-100: rgba(98, 0, 238, 0.025);
+  --purple-alpha-200: rgba(98, 0, 238, 0.05);
+  --purple-200: #E9E3FF;
+  --purple-300: #E7A7FF;
+  --purple-400: #8769FE;
+  --purple-500: #A435F0;
+  --purple-800: #341268;
+  --purple-gradient-200: linear-gradient(180deg, #E9E3FF, white);
+  --purple-gradient-800: : linear-gradient(180deg, #341268, #8769FE);
 
-      --blue-400: #5865F2;
+  --blue-400: #5865F2;
 
-      --red-300: #FF3434;
-      --red-400: #f00;
+  --red-300: #FF3434;
+  --red-400: #f00;
 
-      --green-300: #1AC99F;
+  --green-300: #1AC99F;
 
 }
 
@@ -126,6 +130,7 @@ code {
 ul {
   list-style-type: none;
 }
+
 a {
   text-decoration: none;
   color: unset;
@@ -142,15 +147,19 @@ h3,
 h4 {
   letter-spacing: var(--letter-spacing);
 }
+
 h1 {
   font-size: 3rem;
 }
+
 h2 {
   font-size: 2rem;
 }
+
 h3 {
   font-size: 1.25rem;
 }
+
 h4 {
   font-size: 0.875rem;
 }
@@ -170,18 +179,23 @@ h4 {
   h1 {
     font-size: 4rem;
   }
+
   h2 {
     font-size: 2.5rem;
   }
+
   h3 {
     font-size: 2rem;
   }
+
   h4 {
     font-size: 1.25rem;
   }
+
   body {
     font-size: 1rem;
   }
+
   h1,
   h2,
   h3,
@@ -201,51 +215,51 @@ SOCIAL STYLE
 }
 
 [id$="menu"] {
-    display: none;
-  }
+  display: none;
+}
 
 .show-dropdown {
-    display: inherit !important;
-  }
+  display: inherit !important;
+}
 
-  .rotate-icon {
+.rotate-icon {
   transform: rotateX(180deg) !important;
 }
 
 .sidebar-open {
-    overflow: hidden;
-    height: 100vh;
+  overflow: hidden;
+  height: 100vh;
 }
 
-@media screen and (min-width: 1024px){
+@media screen and (min-width: 1024px) {
   .sidebar-open {
     overflow: auto;
     height: unset;
-}
+  }
 }
 
 
-.brand-text{
+.brand-text {
   color: var(--purple-400)
 }
 
-.brand-background{
+.brand-background {
   background-color: var(--purple-300);
 }
 
-.link-default-style{
+.link-default-style {
   text-decoration: none;
   color: 'unset',
 }
 
-.desktop-new-line{
+.desktop-new-line {
   display: none;
 }
 
-@media screen and (min-width: 767px){
- .desktop-new-line{
-  display: block;
-}
+@media screen and (min-width: 767px) {
+  .desktop-new-line {
+    display: block;
+  }
 }
 
 .rotate-icon {
@@ -259,4 +273,9 @@ SOCIAL STYLE
   -moz-user-select: none;
   -ms-user-select: none;
   user-select: none;
+}
+
+.Tooltip {
+  pointer-events: none;
+
 }
